Updating Turbine to extends ProjectStructure
Moving validation from Turbine into ProjectStructure
